requirements.txt
google-analytics-data>=0.18.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-genai>=1.0.0
notion-client>=2.0.0
python-dotenv>=1.0.0
jinja2>=3.1.0
